add system for dry running setup

+++ setup_tools/symlink.py
@@ -1,6 +1,9 @@
 import os
 
+from setup_tools.config import config
+
 _symlinks = []
+
 
 def add_symlink(src, dest):
     src = os.path.expanduser(src)
@@ -10,12 +13,14 @@ def add_symlink(src, dest):
             print(f'{dest} already exists and is a link')
         else:
             print(f'Error. {dest} already exists')
-        return False
+        return
+
+    if config['dry_run']:
+        return
 
     print(f'{dest} does not exist. Creating symlink...')
     os.makedirs(os.path.dirname(dest), exist_ok=True)
     os.symlink(src, dest)
-    return True
 
 
 def symlink(src, dest):
